requirements.txt
altgraph==0.17.5
lxml==6.0.2
macholib==1.16.4
packaging==25.0
pillow==12.1.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2025.11
PyQt6==6.10.2
PyQt6-Qt6==6.10.1
PyQt6_sip==13.11.0
python-docx==1.2.0
setuptools==80.9.0
typing_extensions==4.15.0
